Separate messages for Default Media Receiver from other message types
