+++ SharpCaster/Models/MessageFactory.cs
@@ -64,50 +64,17 @@ namespace SharpCaster.Models
             PayloadUtf8 = new GetStatusRequest().ToJson()
         };
 
-        public static CastMessage Play(string destinationId, long mediaSessionId) => new CastMessage(destinationId, UniqueSourceID)
-        {
-            Namespace = DialConstants.DialMediaUrn,
-            PayloadUtf8 = new PlayRequest(mediaSessionId).ToJson()
-        };
-
-        public static CastMessage Pause(string destinationId, long mediaSessionId) => new CastMessage(destinationId, UniqueSourceID)
-        {
-            
-            Namespace = DialConstants.DialMediaUrn,
-            PayloadUtf8 = new PauseRequest(mediaSessionId).ToJson()
-        };
-
         public static CastMessage Launch(string appId) => new CastMessage
         {
             Namespace = DialConstants.DialReceiverUrn,
             PayloadUtf8 = new LaunchRequest(appId).ToJson()
         };
 
-        public static CastMessage Load(string destinationId, string payload) => new CastMessage(destinationId, UniqueSourceID)
-        {
-            Namespace = DialConstants.DialMediaUrn,
-            PayloadUtf8 = payload
-        };
-
-        public static CastMessage Seek(string destinationId, long mediaSessionId, double seconds)
-            => new CastMessage(destinationId, UniqueSourceID)
-            {
-                Namespace =  DialConstants.DialMediaUrn,
-                PayloadUtf8 = new SeekRequest(mediaSessionId, seconds).ToJson()
-            };
-
         public static CastMessage Stop(string sessionId) => new CastMessage
         {
             Namespace = DialConstants.DialReceiverUrn,
             PayloadUtf8 = new StopRequest(sessionId).ToJson()
         };
-
-        public static CastMessage MediaStatus(string destinationId) => new CastMessage(destinationId, UniqueSourceID)
-        {
-            Namespace = DialConstants.DialMediaUrn,
-            PayloadUtf8 = new MediaStatusRequest().ToJson()
-        };
-        
     }
 
 }
